Try to deal with the bug on loading/writing SalesRecord lists...
The main issue is that the Item and its child classes (Product and Promotion) actually didn't write correctly to JSON file. It cannot be loaded without a JSON converter anyway. So the problem comes to be more complex.

I recommend that the SalesRecord related code should be refactored in order to keep it clean and neatly.